requirements.txt
amqp==2.4.1
astroid==2.1.0
atomicwrites==1.3.0
attrs==18.2.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
djangorestframework==3.9.1
docker==3.7.0
docker-pycreds==0.4.0
idna==2.8
isort==4.3.4
kombu==4.3.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==5.0.0
pluggy==0.8.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.7.0
pylint-plugin-utils==0.4
pytest==4.2.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
vine==1.2.0
websocket-client==0.54.0
wrapt==1.11.1
